change item id type to long, return from save function, replace on insert conflict

+++ domain/src/main/java/com/ashwinrao/packup/domain/usecase/GetItemUseCase.kt
@@ -9,5 +9,5 @@ import javax.inject.Inject
 class GetItemUseCase
 @Inject
 constructor(private val repo: ItemRepository) {
-    suspend operator fun invoke(id: Int): Item = repo.getItem(id)
+    suspend operator fun invoke(id: Long): Item = repo.getItem(id)
 }
